requirements.txt
# Core Google ADK dependencies
google-adk
litellm
python-dotenv

# Database for sessions
sqlalchemy

# Optional but recommended
uvicorn
fastapi